// frontend/src/types/openapiGenerator/models/GetWeatherRiskCardListResponse.ts
/**
 * Nyeoreumnagi Swagger
 * 녀름나기 REST API 문서
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MyCropDto } from '../models/MyCropDto';
import { Risk } from '../models/Risk';

export class GetWeatherRiskCardListResponse {
    'riskName'?: string;
    'risks'?: Array<Risk>;
    'myCropList'?: Array<MyCropDto>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "riskName",
            "baseName": "riskName",
            "type": "string",
            "format": ""
        },
        {
            "name": "risks",
            "baseName": "risks",
            "type": "Array<Risk>",
            "format": ""
        },
        {
            "name": "myCropList",
            "baseName": "myCropList",
            "type": "Array<MyCropDto>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetWeatherRiskCardListResponse.attributeTypeMap;
    }

    public constructor() {
    }
}