// frontend/src/types/openapiGenerator/models/PestCard.ts
/**
 * Nyeoreumnagi Swagger
 * 녀름나기 REST API 문서
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class PestCard {
    'pestRiskId'?: number;
    'damageType'?: PestCardDamageTypeEnum;
    'name'?: string;
    'description'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "pestRiskId",
            "baseName": "pestRiskId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "damageType",
            "baseName": "damageType",
            "type": "PestCardDamageTypeEnum",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PestCard.attributeTypeMap;
    }

    public constructor() {
    }
}

export const PestCardDamageTypeEnum = {
  Germs: 'GERMS',
  Bugs: 'BUGS'
} as const;
export type PestCardDamageTypeEnum = typeof PestCardDamageTypeEnum[keyof typeof PestCardDamageTypeEnum];